chore: rename type file
